Split SourceKitTests.swift
SourceKitTests.swift was a collection of unrelated tests. Split the apart into multiple, dedicated test files
